src/test/java/ru/sovetnikov/app/web/restaurant/UserRestaurantControllerTest.java
package ru.sovetnikov.app.web.restaurant;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.security.test.context.support.WithUserDetails;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import ru.sovetnikov.app.repository.RestaurantRepository;
import ru.sovetnikov.app.repository.UserRepository;
import ru.sovetnikov.app.web.AbstractControllerTest;
import ru.sovetnikov.app.web.user.AdminUserController;

import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static ru.sovetnikov.app.web.restaurant.RestaurantTestData.NOT_FOUND;
import static ru.sovetnikov.app.web.restaurant.UserRestaurantController.REST_URL;
import static ru.sovetnikov.app.web.restaurant.RestaurantTestData.*;
import static ru.sovetnikov.app.web.user.UserTestData.*;

public class UserRestaurantControllerTest extends AbstractControllerTest {

    @Autowired
    private RestaurantRepository restaurantRepository;
    private static final String REST_URL_SLASH = REST_URL + '/';

    @Test
    @WithUserDetails(value = USER_MAIL)
    void getAll() throws Exception {
        perform(MockMvcRequestBuilders.get(REST_URL))
                .andExpect(status().isOk())
                .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
                .andExpect(RestaurantTestData.RESTAURANT_TO_MATCHER.contentJson(restaurantsTo));
    }

    @Test
    @WithUserDetails(value = USER_MAIL)
    void getOne() throws Exception {
        perform(MockMvcRequestBuilders.get(REST_URL_SLASH + RESTAURANT1_ID))
                .andExpect(status().isOk())
                .andDo(print())
                .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
                .andExpect(RESTAURANT_TO_WITH_MENU_MATCHER.contentJson(RESTAURANT1_TO));
    }

    @Test
    void getUnauthorized() throws Exception{
        perform(MockMvcRequestBuilders.get(REST_URL))
                .andExpect(status().isUnauthorized());
    }

    @Test
    @WithUserDetails(value = USER_MAIL)
    void getNotFound() throws Exception {
        perform(MockMvcRequestBuilders.get(REST_URL + NOT_FOUND))
                .andDo(print())
                .andExpect(status().isNotFound());
    }
}
